+++ src/app/showcase/page.tsx
@@ -1,80 +1,170 @@
 'use client'
-import { defaultGetFetcher } from '@/api/fetchers'
 
-import { FormControlLabel, Stack, Typography } from '@mui/material'
-import React, { useEffect, useState } from 'react'
-import useSWR from 'swr'
-import { IPagination } from '@/types/IPagination'
+import { IconButton, Stack, Typography } from '@mui/material'
+import ShowcaseCard from './panel/ShowcaseCard'
+import { TouchEvent, useCallback, useEffect, useRef, useState } from 'react'
+import { motion } from 'framer-motion'
+import PhoneFrame from './panel/PhoneFrame'
 import useMedia from '@/hook/useMedia'
-import CardContainer from './panel_refac/CardContainer'
+import ExpandLessIcon from '@mui/icons-material/ExpandLess'
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
+import useSWR from 'swr'
 import { ICardData } from './panel/types'
+import { defaultGetFetcher } from '@/api/fetchers'
+import { IPagination } from '@/types/IPagination'
 import useAuthStore from '@/states/useAuthStore'
 import useAxiosWithAuth from '@/api/config'
-import { AxiosInstance } from 'axios'
 
-// 1. 메인페이지 api 긁어오기
-// 2. 메인 카드 컴포넌트 가져오기
-// 3. 틴더 어쩌구 적용
-
-const Showcase = () => {
+const ShowcasePage = () => {
   const [page, setPage] = useState<number>(1)
-  const { isLogin } = useAuthStore()
-  const [cardList, setCardList] = useState<Array<ICardData>>([])
-  const axiosInstance: AxiosInstance = useAxiosWithAuth()
+  const [cards, setCards] = useState<ICardData[]>([])
+  const [index, setIndex] = useState(0)
+  const [touchStart, setTouchStart] = useState(0)
+  const [touchEnd, setTouchEnd] = useState(0)
+  const constraintsRef = useRef(null)
   const { isPc } = useMedia()
+  const { isLogin } = useAuthStore()
+  const axiosWithAuth = useAxiosWithAuth()
   const { data, error, isLoading } = useSWR<IPagination<ICardData[]>>(
     `${process.env.NEXT_PUBLIC_API_URL}/api/v1/showcase?page=${page}&pageSize=10`,
     isLogin
-      ? (url: string) => axiosInstance.get(url).then((res) => res.data)
+      ? (url: string) => axiosWithAuth.get(url).then((res) => res.data)
       : defaultGetFetcher,
   )
 
-  console.log(data)
-
   useEffect(() => {
     if (!isLoading && data?.content) {
-      setCardList((prev) => {
-        const newArray = [...data.content].reverse().concat(prev)
-        return newArray
-      })
+      // setCards((prev) => [...data.content].reverse().concat(prev))
+      if (data.first) setCards([...data.content])
+      else setCards([...cards, ...data.content])
     }
-  }, [isLoading, data?.content])
+  }, [isLoading, data, setCards])
 
-  let message: string = ''
+  const handlePrevClick = () => {
+    if (!data) return
+    if (index > 0) {
+      setIndex(index - 1)
+    }
+  }
 
-  if (isLoading && !cardList.length) message = '로딩중'
-  else if (error) message = '에러 발생'
+  const handleNextClick = useCallback(() => {
+    if (!cards) return
+    if (index < cards.length - 1) {
+      setIndex(index + 1)
+      if (index === cards.length - 3) {
+        setPage(page + 1)
+      }
+    }
+  }, [index, page, cards])
+
+  const handleTouchStart = (e: TouchEvent) => {
+    e.stopPropagation()
+    setTouchStart(e.targetTouches[0].clientY)
+  }
+
+  const handleTouchMove = (e: TouchEvent) => {
+    e.stopPropagation()
+    setTouchEnd(e.targetTouches[0].clientY)
+  }
+
+  const handleTouchEnd = (e: TouchEvent) => {
+    e.stopPropagation()
+    if (touchEnd < touchStart + 200) {
+      if (!data) return
+      if (index < data.content.length - 1) {
+        setIndex(index + 1)
+        if (index === cards.length - 3) {
+          setPage(page + 1)
+        }
+        setTouchEnd(0)
+        setTouchStart(0)
+      }
+    } else {
+      if (!data) return
+      if (index > 0) {
+        setIndex(index - 1)
+        setTouchEnd(0)
+        setTouchStart(0)
+      }
+    }
+  }
+
+  const variants = {
+    hidden: { opacity: 0, y: 100 },
+    visible: { opacity: 1, y: 0 },
+    exit: { opacity: 0, y: -100 },
+  }
+
+  var message = ''
+
+  if (error) message = '에러 발생'
+  else if (!data) message = '로딩 중...'
+  else if (data.content.length === 0) message = '데이터가 없습니다'
+
+  if (isPc) {
+    return (
+      <>
+        <Stack direction={'row'} spacing={2} height={'600px'}>
+          {message ? (
+            <Typography>{message}</Typography>
+          ) : (
+            <>
+              <Stack direction={'row'} spacing={1}>
+                <PhoneFrame
+                  imageUrl={cards[index] ? cards[index].image! : undefined}
+                />
+                <Stack direction={'column-reverse'}>
+                  <Stack>
+                    <IconButton onClick={handlePrevClick}>
+                      <ExpandLessIcon color="primary" />
+                    </IconButton>
+                    <IconButton onClick={handleNextClick}>
+                      <ExpandMoreIcon color="primary" />
+                    </IconButton>
+                  </Stack>
+                </Stack>
+              </Stack>
+              <ShowcaseCard data={cards ? cards[index] : undefined} />
+            </>
+          )}
+        </Stack>
+      </>
+    )
+  }
 
   return (
     <Stack
-      justifyContent={'space-between'}
-      alignItems={'center'}
-      sx={{ width: '100%', height: '80svh' }}
-      direction={'column'}
+      component={'div'}
+      onTouchStart={handleTouchStart}
+      onTouchMove={handleTouchMove}
+      onTouchEnd={handleTouchEnd}
     >
-      <Stack
-        justifyContent={'center'}
-        alignItems={'center'}
-        sx={{
-          width: isPc ? '20.5rem' : '92vw',
-          height: '27rem',
-          maxWidth: '20.5rem',
-          position: 'relative',
-        }}
-      >
-        {!message ? (
-          <CardContainer
-            cardList={cardList}
-            isLoading={isLoading}
-            update={() => setPage((prev) => (!data?.last ? prev + 1 : prev))}
-            setCardList={setCardList}
-          />
-        ) : (
-          <Typography>{message}</Typography>
-        )}
-      </Stack>
+      {isLoading && page === 1 ? (
+        <Typography>로딩 중...</Typography>
+      ) : error || !data ? (
+        <Typography>에러 발생</Typography>
+      ) : data.content.length === 0 ? (
+        <Typography>데이터가 없습니다</Typography>
+      ) : (
+        <motion.div ref={constraintsRef}>
+          <motion.div
+            key={index}
+            variants={variants}
+            drag="y"
+            dragConstraints={constraintsRef}
+            dragTransition={{ bounceStiffness: 600, bounceDamping: 20 }}
+            dragDirectionLock
+            initial="hidden"
+            animate="visible"
+            exit={'exit'}
+            transition={{ duration: 0.3 }}
+          >
+            <ShowcaseCard data={cards[index]} />
+          </motion.div>
+        </motion.div>
+      )}
     </Stack>
   )
 }
 
-export default Showcase
+export default ShowcasePage
